# docs/arc-task-ledger/settlement-verifier/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arc USDC 结算核验 API / Settlement Verifier",
    "version": "2.0.0",
    "description": "Read-only chain access. Atomic18 string units. Offline integrity is not mainnet authentication. Existing legacy API contract remains separate."
  },
  "servers": [
    {
      "url": "https://web--atl-web--xtd599t97njk.code.run/api"
    },
    {
      "url": "http://127.0.0.1:8095"
    }
  ],
  "paths": {
    "/v1/verifier": {
      "get": {
        "summary": "版本、注册网络与有界资源 / versions and limits",
        "description": "Reads existing state only; no RPC work.",
        "responses": {
          "200": {
            "description": "Result; HTTP 200 does not imply MATCHED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verifier/examples": {
      "get": {
        "summary": "最多三份明确公开的固定报告，无种子 / up to three persisted public examples",
        "description": "Reads existing state only; no RPC work.",
        "responses": {
          "200": {
            "description": "Result; HTTP 200 does not imply MATCHED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions": {
      "post": {
        "summary": "建立或复用7天会话 / issue or reuse session",
        "description": "Only bounded service work/storage. Chain access remains read-only. Browser Origin/CSRF or SDK X-Arc-Client are required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "maxProperties": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result; HTTP 200 does not imply MATCHED",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "csrfToken": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "sessionToken": {
                      "type": "string",
                      "description": "SDK only; do not publish"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verifications": {
      "post": {
        "summary": "单笔主网核验并持久保存 / bounded transaction verification",
        "description": "Only bounded service work/storage. Chain access remains read-only. Browser Origin/CSRF or SDK X-Arc-Client are required.",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,100}$"
            }
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result; HTTP 200 does not imply MATCHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "202": {
            "description": "Same key still RUNNING. Read status; no second acquisition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "const": "VERIFICATION_RUNNING"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/verifications/{id}": {
      "get": {
        "summary": "所有者读取持久请求状态 / owner request status",
        "description": "Reads existing state only; no RPC work.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Result; HTTP 200 does not imply MATCHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            }
          },
          "400": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "明确输入、身份、版本、配额或来源错误；不能当成付款成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zasv_session"
      },
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Verification": {
        "type": "object",
        "additionalProperties": false,
        "required": ["transaction"],
        "properties": {
          "transaction": {
            "type": "string",
            "maxLength": 200,
            "description": "Canonical transaction hash or registered explorer.arc.io/tx/HASH; parsed only, never fetched"
          },
          "expectation": {
            "$ref": "#/components/schemas/Expectation"
          },
          "task": {
            "type": "object",
            "additionalProperties": false,
            "required": ["jobId", "snapshotRunId", "legId"],
            "properties": {
              "jobId": {
                "type": "string",
                "pattern": "^[0-9]{1,78}$"
              },
              "snapshotRunId": {
                "type": "string",
                "maxLength": 100
              },
              "legId": {
                "type": "string",
                "maxLength": 200
              }
            }
          }
        }
      },
...
      "Document": {
        "type": "object",
        "required": [
          "report",
          "observation",
          "evaluation",
          "evidence",
          "snapshot",
          "coverage",
          "freshness",
          "confidence",
          "formalForensicMode",
          "readOnlyChain"
        ],
        "properties": {
          "report": {
            "$ref": "#/components/schemas/Report"
          },
          "observation": {
            "type": "object"
          },
          "evaluation": {
            "type": ["object", "null"]
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "evidence": {
            "type": "object"
          },
          "snapshot": {
            "type": "object"
          },
          "coverage": {
            "type": "object"
          },
          "freshness": {
            "type": "object"
          },
          "confidence": {
            "type": "object"
          },
          "formalForensicMode": {
            "const": false
          },
          "readOnlyChain": {
            "const": true
          }
        }
      },
...
    }
  }
}